+++ src/main/java/com/BEJ/Bej/dto/response/productResponse/ProductAttributeResponse.java
@@ -14,9 +14,9 @@ public class ProductAttributeResponse {
     String name;
 //    String value;
 
-    int originalPrice;
-    int finalPrice;
-    int discount;
+    double originalPrice;
+    double finalPrice;
+    double discount;
 
     int stockQuantity;
     int soldQuantity;
