+++ src/main/java/com/BEJ/Bej/entity/product/ProductAttribute.java
@@ -23,9 +23,9 @@ public class ProductAttribute {
     String name;
 //    String value;
 
-    int originalPrice;
-    int finalPrice;
-    int discount = 0;
+    double originalPrice;
+    double finalPrice;
+    double discount = 0;
 
     int stockQuantity = 0;
     int soldQuantity = 0;
